Use C++14 move-capture lambda in deriving_apply.
The previous commit changed deriving_apply<in_terms_of_bind<M>> to accept
rvalues and references instead of values only (required for the monad version
of std::unique_ptr which cannot be coppied). This caused a segfault in lazy's
test of applicative::apply due to (most likely) a dangling reference in the
rvalue overload.

Capture by move to so that there is no reference to dangle.