Enable manual overload of mem functions/functors.
Implement _dtl::Forwarder<Ret (Obj::*)(Args...)> and <Mem Obj::*> to cover
member functions and member function objects. Since member function objects
cannot be deduced from regular members, it may not be possible to write a
version that can return the member if not a function, and the result of calling
if it is.